bluetooth: tester: Move core service to separate file
There is no need to keep it in same file as BTP protocol handler.
